backend/pyproject.toml
[project]
name = "rag-embeddings-pipeline"
version = "0.1.0"
description = "RAG Embeddings Pipeline for Physical AI & Humanoid Robotics textbook"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "cohere>=4.0.0",
    "qdrant-client>=1.7.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "lxml>=4.9.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.5.0",
    "openai>=1.12.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "slowapi>=0.1.9",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
]